modify the regularity cls

+++ rgr/constants/types.pyx
@@ -1,7 +1,7 @@
 import numpy as np
 
 DTYPE_ADJ = np.int8
-DTYPE_STD = np.int32
+DTYPE_STD = np.int16
 DTYPE_FLOAT = np.float32
 
 cpdef type get_dtype_adj():
